% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era.R
\docType{data}
\name{era}
\alias{era}
\title{ERA5-Land macroclimatic data}
\format{
A data frame with X rows and X variables: \describe{
\item{time}{description} \item{lat}{description} \item{lon}{description}
\item{tas}{description}}
}
\usage{
era
}
\description{
From
https://developers.google.com/earth-engine/datasets/catalog/ECMWF_ERA5_LAND_HOURLY?hl=fr, #nolint
to be described.
}
\keyword{datasets}
