% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_amplitude}
\alias{fft_amplitude}
\title{FFT amplitude}
\usage{
fft_amplitude(coefficients)
}
\arguments{
\item{coefficients}{num. The Fourier coefficients.}
}
\value{
\eqn{a_{n}} array of amplitudes.
}
\description{
Fast Fourier Transform (FFT): get amplitudes from coefficients
}
\examples{

fft_rfft(temperatures = hobo$t_hobo[1:(24 * 5)]) |>
  fft_amplitude()

}
