% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_coefficient}
\alias{fft_coefficient}
\title{FFT Rebuild coefficient}
\usage{
fft_coefficient(amplitudes, phases)
}
\arguments{
\item{amplitudes}{num. The Fourier amplitudes}

\item{phases}{num. The Fourier phases}
}
\value{
\eqn{c} the Fourier coefficients of \eqn{S} of length \eqn{\ell}, whose
coefficients are \eqn{c_{0},c_{1},\ldots,c_{\ell-1}}
}
\description{
Fast Fourier Transform (FFT): get coefficients from amplitude and phase
}
\examples{

c <- fft_rfft(temperatures = hobo$t_hobo[1:(24 * 5)])
a <- fft_amplitude(c)
p <- fft_phase(c)
fft_coefficient(amplitudes = a, phases = p)

}
