% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_delay}
\alias{fft_delay}
\title{FFT delay}
\usage{
fft_delay(f, freq)
}
\arguments{
\item{f}{num. The Fourier coefficients.}

\item{freq}{num. The harmonics frequencies.}
}
\value{
\eqn{D} array of delays, \eqn{D[n] =\theta_{n}} for \eqn{n=1,\ldots,\ell-1}.
}
\description{
Fast Fourier Transform (FFT): delay per frequencies
}
\examples{

fft_rfft(hobo$t_hobo[1:(24 * 5)]) |>
  fft_delay(fft_freq(24 * 5, 24 * 5))

}
