% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_energy}
\alias{fft_energy}
\title{FFT energy}
\usage{
fft_energy(f)
}
\arguments{
\item{f}{num. The Fourier coefficients.}
}
\value{
\eqn{E} the energy.
}
\description{
Fast Fourier Transform (FFT): energy
}
\details{
On the other hand, the so-called Parseval formula gives the \emph{energy}:
\deqn{E = |c_{0}|^{2} + \frac 1 2\sum_{n=1}^{\ell-1} |c_{n}|^{2} = \frac 1 T
\int_{0}^{T} |s(t)|^{2}dt \simeq \frac 1 N \sum_{n=1}^{N} |S_{n}|^{2}}
}
\examples{

fft_rfft(hobo$t_hobo[1:(24 * 5)]) |>
  fft_energy()

}
