% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_mean}
\alias{fft_mean}
\title{FFT mean}
\usage{
fft_mean(coefficients)
}
\arguments{
\item{coefficients}{num. The Fourier amplitudes.}
}
\value{
\eqn{f[1]} the first real component of \eqn{f}.
}
\description{
Fast Fourier Transform (FFT): mean temperature
}
\details{
The amplitude \eqn{a_{0}} plays a particular role, it is the mean
\deqn{a_{0} = \frac 1 N \sum_{n=1}^{N}S_{n}\,.}
}
\examples{

fft_rfft(temperatures = hobo$t_hobo[1:(24 * 5)]) |>
  fft_mean()

}
