% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_period}
\alias{fft_period}
\title{FFT periods}
\usage{
fft_period(frequencies)
}
\arguments{
\item{frequencies}{num. The harmonics frequencies.}
}
\value{
\eqn{p} the harmonics periods, \eqn{p=1/f}.
}
\description{
Fast Fourier Transform (FFT):  periods for a real signal of length T (in
time) with n samples.
}
\examples{

fft_freq(samples = 201, time_window = 10) |>
  fft_period()

}
