% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_phase}
\alias{fft_phase}
\title{FFT phase}
\usage{
fft_phase(coefficients)
}
\arguments{
\item{coefficients}{num. The Fourier coefficients.}
}
\value{
\eqn{b_{n}} array of phases.
}
\description{
Fast Fourier Transform (FFT): get phases from coefficients
}
\examples{

fft_rfft(temperatures = hobo$t_hobo[1:(24 * 5)]) |>
  fft_phase()

}
