% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_power}
\alias{fft_power}
\title{FFT power}
\usage{
fft_power(coefficients)
}
\arguments{
\item{coefficients}{num. The Fourier coefficients.}
}
\value{
\eqn{P} array of powers, \eqn{P[n] = |c_{n}|} for \eqn{n=1,\ldots,\ell-1}.
}
\description{
Fast Fourier Transform (FFT): powers
}
\details{
The power of the \eqn{n}-th harmonics is
\deqn{|c_{n}| = \sqrt{a_{n}^{2}+b_{n}^{2}}\,,}
its frequency is \eqn{n/T} and period \eqn{T/n}, for \eqn{n=1,\ldots,\ell-1}.
}
\examples{

fft_rfft(temperatures = hobo$t_hobo[1:(24 * 5)]) |>
  fft_power()

}
