% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_ratio}
\alias{fft_ratio}
\title{Fourier ratio}
\usage{
fft_ratio(
  data,
  t,
  index_col,
  microclimate_col,
  macroclimate_col,
  window = "5 days",
  step = "3 days",
  period = TRUE,
  power = TRUE
)
}
\arguments{
\item{data}{df. The data frame with a column containing the time index and a
column containing the temperature.}

\item{t}{int. Time window, \eqn{t=N\Delta t} where \eqn{\Delta t} is the
sampling interval.}

\item{index_col}{char. The name of the column containing the time index.}

\item{microclimate_col}{char. The name of the column containing the
microclimate temperature series.}

\item{macroclimate_col}{char. The name of the column containing the
macroclimate temperature series.}

\item{window}{time. Window size, can be expressed in time units see
\link[runner:runner]{runner::runner}, default is 5 days}

\item{step}{time. Window step,  can be expressed in time units see
\link[runner:runner]{runner::runner}, defaults is 3 days.}

\item{period}{bool. To include period or not, default TRUE.}

\item{power}{bool. To include power or not, default TRUE.}
}
\value{
A table with:
\itemize{
\item the datetime of the window
\item \eqn{f} the harmonics frequencies \eqn{1/y,\ldots \ell/t}
\item \eqn{p} the harmonics periods
\item \eqn{f} the Fourier coefficients of \eqn{s} of length \eqn{\ell}, whose
coefficients are \eqn{c_{0},c_{1},\ldots,c_{\ell-1}} for micro and
macroclimate.
\item \eqn{P} array of powers, \eqn{P[n] = |c_{n}|} for \eqn{n=1,\ldots,\ell-1}
for micro and macroclimate.
\item \eqn{R} array of ratios of micro vs. macro-climate powers
}
}
\description{
Fast Fourier Transform (FFT) of two series of microclimatic and macroclimatic
temperatures in a table with a rolling window across a time index with
resulting frequencies, periods, micro and macro coefficients and micro and
macro powers, and the ratio of power between the micro and macroclimate per
window in a single table.
}
\details{
Can be used with a grouped table, see vignette \emph{to be linked}.
}
\examples{

data <- era |>
  dplyr::rename(era = tas, datetime = time) |>
  dplyr::select(datetime, era) |>
  dplyr::left_join(dplyr::select(hobo, datetime, t_hobo) |>
    dplyr::rename(hobo = t_hobo))
fft_ratio(data, 24 * 5, "datetime", "hobo", "era") |>
  dplyr::filter(period == 24) |>
  summary()

}
