% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_reconstruct}
\alias{fft_reconstruct}
\title{FFT reconstruct}
\usage{
fft_reconstruct(coefficients, frequencies, times)
}
\arguments{
\item{coefficients}{num. The Fourier coefficients.}

\item{frequencies}{num. The harmonics frequencies.}

\item{times}{num. Time array.}
}
\value{
\eqn{s} an array that contains the value of the series of temperature
\eqn{s(t)}.
}
\description{
Fast Fourier Transform (FFT): Reconstruct the temperatures at time t, whose
coefficients are 'c' associate to frequencies 'f'
}
\examples{

fft_reconstruct(
  coefficients = fft_rfft(temperatures = hobo$t_hobo[1:(24 * 5)]),
  frequencies = fft_freq(samples = 24 * 5, time_window = 24 * 5),
  times = 1:24 * 5
)

}
