% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_roll}
\alias{fft_roll}
\title{Fourier roll}
\usage{
fft_roll(
  data,
  time_window,
  time_col,
  temperature_col,
  window = "5 days",
  step = "3 days",
  period = TRUE,
  power = TRUE
)
}
\arguments{
\item{data}{df. The data frame with a column containing the time index and a
column containing the temperature.}

\item{time_window}{int. Time window, \eqn{T=n\Delta t} where \eqn{\Delta t}
is the sampling interval.}

\item{time_col}{char. The name of the column containing the time index.}

\item{temperature_col}{char. The name of the column containing the
temperature series.}

\item{window}{time. Window size, can be expressed in time units see
\link[runner:runner]{runner::runner}, default is 5 days}

\item{step}{time. Window step,  can be expressed in time units see
\link[runner:runner]{runner::runner}, defaults is 3 days.}

\item{period}{bool. To include period or not, default TRUE.}

\item{power}{bool. To include power or not, default TRUE.}
}
\value{
A table with:
\itemize{
\item the datetime of the window
\item \eqn{f} the harmonics frequencies \eqn{1/y,\ldots \ell/t}
\item \eqn{p} the harmonics periods
\item \eqn{a_{n}} the amplitudes of the Fourier coefficients of \eqn{s}
of length \eqn{\ell}, whose values are \eqn{a_{0},a_{1},\ldots,c_{\ell-1}}
\item \eqn{b_{n}} the phases of the Fourier coefficients of \eqn{s}
of length \eqn{\ell}, whose values are \eqn{a_{0},a_{1},\ldots,c_{\ell-1}}
\item \eqn{P} array of powers, \eqn{P[n] = |c_{n}|} for \eqn{n=1,\ldots,\ell-1}
}
}
\description{
Fast Fourier Transform (FFT) of a series of temperatures in a table with a
rolling window across a time index with resulting frequencies, periods,
coefficients and powers per window in a single table.
}
\details{
Can be used with a grouped table, see vignette \emph{to be linked}.
}
\examples{

fft_roll(
  data = hobo, time_window = 24 * 5,
  time_col = "datetime", temperature_col = "t_hobo"
)

}
