% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo.R
\docType{data}
\name{hobo}
\alias{hobo}
\title{HOBO micrloclimatic data}
\format{
A data frame with X rows and X variables: \describe{
\item{id_sensor}{description} \item{id_plot}{description}
\item{position_sensor}{description} \item{datetime}{description}
\item{month}{description} \item{month_num}{description}
\item{date}{description} \item{time}{description}
\item{t_hobo}{description} }
}
\usage{
hobo
}
\description{
From
https://data.indores.fr/dataset.xhtml?persistentId=doi:10.48579/PRO/ZQBQUW,
to be described.
}
\keyword{datasets}
