% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_freq}
\alias{fft_freq}
\title{FFT frequency}
\usage{
fft_freq(samples, time_window)
}
\arguments{
\item{samples}{int. Number of samples.}

\item{time_window}{int. Time window, \eqn{T=n\Delta t} where \eqn{\Delta t}
is the sampling interval.}
}
\value{
\eqn{f} the harmonics frequencies \eqn{1/t,\ldots (\ell-1)/t}
}
\description{
Fast Fourier Transform (FFT):  frequencies for a real signal of length T (in
time) with n samples.
}
\examples{

fft_freq(samples = 201, time_window = 10)

}
