% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_base.R
\name{fft_freq}
\alias{fft_freq}
\title{FFT frequencies}
\usage{
fft_freq(n, t)
}
\arguments{
\item{n}{int. Number of samples.}

\item{t}{int. Time window, \eqn{t=N\Delta t} where \eqn{\Delta t} is the
sampling interval.}
}
\value{
\eqn{f} the harmonics frequencies \eqn{1/y,\ldots \ell/t}.
}
\description{
Fast Fourier Transform (FFT):  frequencies for a real signal of length t (in
time) with n samples.
}
\examples{

fft_freq(201, 10)

}
