% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_ratio}
\alias{fft_ratio}
\title{Fourier ratio}
\usage{
fft_ratio(
  data,
  time_window,
  time_col,
  microclimate_col,
  macroclimate_col,
  window = "5 days",
  step = "3 days",
  period = TRUE,
  power = TRUE
)
}
\arguments{
\item{data}{df. The data frame with a column containing the time index and a
column containing the temperature.}

\item{time_window}{int. Time window, \eqn{T=N\Delta t} where \eqn{\Delta t}
is the sampling interval.}

\item{time_col}{char. The name of the column containing the time index.}

\item{microclimate_col}{char. The name of the column containing the
microclimate temperature series.}

\item{macroclimate_col}{char. The name of the column containing the
macroclimate temperature series.}

\item{window}{time. Window size, can be expressed in time units see
\link[runner:runner]{runner::runner}, default is 5 days}

\item{step}{time. Window step,  can be expressed in time units see
\link[runner:runner]{runner::runner}, defaults is 3 days.}

\item{period}{bool. To include period or not, default TRUE.}

\item{power}{bool. To include power or not, default TRUE.}
}
\value{
A table with:
\itemize{
\item the datetime of the window
\item \eqn{f} the harmonics frequencies \eqn{1/y,\ldots \ell/t}
\item \eqn{p} the harmonics periods
\item \eqn{a_{n}} the amplitudes of the Fourier coefficients of \eqn{s}
of length \eqn{\ell}, whose values are \eqn{a_{0},a_{1},\ldots,c_{\ell-1}}
\item \eqn{b_{n}} the phases of the Fourier coefficients of \eqn{s}
of length \eqn{\ell}, whose values are \eqn{a_{0},a_{1},\ldots,c_{\ell-1}}
\item \eqn{P} array of powers, \eqn{P[n] = |c_{n}|} for \eqn{n=1,\ldots,\ell-1}
for micro and macroclimate.
\item \eqn{R} array of ratios of micro vs. macro-climate powers
}
}
\description{
Fast Fourier Transform (FFT) of two series of microclimatic and macroclimatic
temperatures in a table with a rolling window across a time index with
resulting frequencies, periods, micro and macro coefficients and micro and
macro powers, and the ratio of power between the micro and macroclimate per
window in a single table.
}
\details{
Can be used with a grouped table, see \href{Examples.html}{vignette}.
}
\examples{

data <- era |>
  dplyr::rename(era = tas, datetime = time) |>
  dplyr::select(datetime, era) |>
  dplyr::left_join(
    dplyr::select(hobo, datetime, t_hobo) |>
      dplyr::rename(hobo = t_hobo),
    by = dplyr::join_by(datetime)
  )
fft_ratio(
  data = data, time_window = 24 * 5, time_col = "datetime",
  microclimate_col = "hobo", macroclimate_col = "era"
) |>
  dplyr::filter(period == 24) |>
  dplyr::summarise_all(mean, na.rm = TRUE)

}
