% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_reconstruct}
\alias{fft_reconstruct}
\title{FFT reconstruct}
\usage{
fft_reconstruct(f, freq, time)
}
\arguments{
\item{f}{num. The Fourier coefficients.}

\item{freq}{num. The harmonics frequencies.}

\item{time}{num. Time array.}
}
\value{
\eqn{s} an array that contains the value of \eqn{s(t)}.
}
\description{
Fast Fourier Transform (FFT): Reconstruct the temperatures at time t, whose
spectrum is 'f' associate to frequencies 'freq'
}
\examples{

fft_reconstruct(
  fft_rfft(hobo$t_hobo[1:(24 * 5)]),
  fft_freq(24 * 5, 24 * 5),
  1:24 * 5
)

}
