% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_base.R
\name{fft_rfft}
\alias{fft_rfft}
\title{Fourier coefficients}
\usage{
fft_rfft(s)
}
\arguments{
\item{s}{num. Sampled temperatures at constant interval.}
}
\value{
\eqn{f} the Fourier coefficients of \eqn{s} of length \eqn{\ell}, whose
coefficients are \eqn{c_{0},c_{1},\ldots,c_{\ell-1}}
}
\description{
Fast Fourier Transform (FFT) of a series of temperatures.
}
\details{
A series \eqn{S} consisting of \eqn{N} samples measured at a time interval
\eqn{\Delta t} can be analyzed by a Fourier series

\deqn{s(t) = a_{0} +  \sum_{n=1}^{\ell} a_{n} \cos(2\pi \tfrac{n}{T} t) +
b_{n}\sin(2\pi \tfrac n T t)}

where \eqn{a_n} and \eqn{b_n} are the Fourier coefficients obtained from the
fast fourier transform (FFT) of \eqn{S}. Note, the time reference of the
first measure is \eqn{t=0}.

The number of coeffcients is
\deqn{
\ell =
\begin{cases}
N/2 + 1 & \text{if } N \text{ is even} \\
(N+1)/2 & \text{if } N \text{ is odd}
\end{cases}
}

Generally they take the form of complex number \deqn{c_{n} = a_{n} - i b_{n}
\,, \quad n=0,\ldots, \ell-1} with the convention \eqn{b_{0}=0}. Hence, we
have the relation \deqn{a_{n} = \Re (c_{n}) \text{ and }  b_{n} = -
\Im(c_{n})\,, \quad n=0,\ldots,\ell \,.}
}
\examples{

fft_rfft(hobo$t_hobo[1:(24*5)])

}
