% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_tab}
\alias{fft_tab}
\title{Fourier table}
\usage{
fft_tab(s, t, period = TRUE, power = TRUE)
}
\arguments{
\item{s}{num. Sampled temperatures at constant interval.}

\item{t}{int. Time window, \eqn{t=N\Delta t} where \eqn{\Delta t} is the
sampling interval.}

\item{period}{bool. To include period or not, default TRUE.}

\item{power}{bool. To include power or not, default TRUE.}
}
\value{
A table with:
\itemize{
\item \eqn{f} the harmonics frequencies \eqn{1/t,\ldots \ell/t}
\item \eqn{p} the harmonics periods
\item \eqn{c} the Fourier coefficients of \eqn{s} of length \eqn{\ell}, whose
coefficients are \eqn{c_{0},c_{1},\ldots,c_{\ell-1}}
\item \eqn{P} array of powers, \eqn{P[n] = |c_{n}|} for \eqn{n=1,\ldots,\ell-1}
}
}
\description{
Fast Fourier Transform (FFT) of a series of temperatures with resulting
frequencies, periods, coefficients and powers in a single table.
}
\examples{

fft_tab(hobo$t_hobo[1:(24 * 5)], 24 * 5)

}
