% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_base.R
\name{fft_variance}
\alias{fft_variance}
\title{FFT variance}
\usage{
fft_variance(f)
}
\arguments{
\item{f}{num. The Fourier coefficients.}
}
\value{
\eqn{V} the variance.
}
\description{
Fast Fourier Transform (FFT): variance
}
\details{
interpreted as a variance by the formula \deqn{\frac 1 2 \sum_{n=1}^{\ell-1}
|c_{n}|^{2} = \frac 1 T \int_{0}^{T} |s(t)-c_{0}|^{2}dt \simeq \frac 1 N
\sum_{n=1}^{N} |S_{n}-c_{0}|^{2}}
}
\examples{

fft_rfft(hobo$t_hobo[1:(24 * 5)]) \%>\%
  fft_variance()

}
